PDFBOX-2147: Removed unneccesary temp during conversion to/from string and double boxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1609330 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/common/PDNumberTreeNode.java
@@ -269,8 +269,8 @@ public class PDNumberTreeNode implements COSObjectable
             Integer upper = null;
             if( keys.size() > 0 )
             {
-                lower = (Integer)keys.get( 0 );
-                upper = (Integer)keys.get( keys.size()-1 );
+                lower = keys.get( 0 );
+                upper = keys.get( keys.size()-1 );
             }
             setUpperLimit( upper );
             setLowerLimit( lower );
